edit no#1 - (program 2)
limited the range of newX and newY to furthur contain the ellipse
placement on the screen. This causes the bolder section of ellipse
placement towards the bottom right of the screen, this happens because x
and y values range to 5000 (which is off the screen) and gets moved back
onto the screen.
I coded the fill to change at every i=1000 interval to illustrate how
the pattern develops over time.